ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
opencv-python==4.10.0.84
numpy==2.1.1
requests==2.32.3
python-dotenv==1.0.1

